refactor: (#25) 웹툰 뷰어 훅 분리
